content/media/video/test/test_seek7.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Media test: seek test 7</title>
  <script type="text/javascript" src="/MochiKit/MochiKit.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<video id='v'
       src='seek.ogg'
       onloadedmetadata='return startTest();'
       onseeked='return seekEnded();'></video>
<pre id="test">
<script class="testbody" type="text/javascript">
// Test for bug identified by Chris Pearce in comment 40 on
// bug 449159.
var v = document.getElementById('v');
var seekCount = 0;
var interval;
var timeout;

function startTest() {
  interval = setInterval(function() { v.currentTime=Math.random*3; }, 10);
}

function seekEnded() {
  seekCount++;
  ok(true, "Seek " + seekCount);
  if (seekCount == 3) {
    clearInterval(interval);
    clearTimeout(timeout);
    SimpleTest.finish();
  }
  return false;
}

timeout = setTimeout(function () {
                       ok(false, "Test timed out");
                       SimpleTest.finish();
                     }, 30000);

SimpleTest.waitForExplicitFinish();
</script>
</pre>
</body>
</html>
